Use std::optional for seed setting

+++ include/Constants.hpp
@@ -5,7 +5,7 @@ namespace DefaultConstants{
     constexpr ProcessType process = ProcessType::BM;
     constexpr double PI = std::numbers::pi_v<double>;
     constexpr double pdfThreshold = 1e-6;
-    constexpr double pdfPoints = 1e3;
+    constexpr size_t pdfPoints = static_cast<size_t>(1e3);
     constexpr size_t maxPathsToDraw = 20;
     constexpr double OUthetaData = 1.0; // Mean-reversion coefficient for OU process (Hard coded for now...)
     namespace Simulation {
